+++ src/Modules/Workflows/Domain/LegacyAction/TriggerWorkflow.php
@@ -44,6 +44,10 @@ class TriggerWorkflow implements ExpirationActionInterface
         $args = $actionsArgsModel->getArgs();
         $args['postId'] = $postId;
 
+        if (isset($args['extraData']['workflowId'])) {
+            $args['workflowId'] = $args['extraData']['workflowId'];
+        }
+
         if (isset($args['expireType']) && $args['expireType'] === self::ACTION_NAME) {
             $this->hooks->doAction(HooksAbstract::ACTION_LEGACY_ACTION, $postId, $post, $args);
         }